cmd/doctree: Move utils functions to utils.go
